requirements.txt
django-crum==0.7.2
diff-match-patch==20121119
dj-database-url==0.4.2
Django==1.11.7
django-import-export==0.5.1
django-tables2==1.14.2
et-xmlfile==1.0.1
jdcal==1.3
odfpy==1.3.5
openpyxl==2.4.9
psycopg2==2.6.2
python-decouple==3.1
pytz==2017.3
PyYAML==3.12
tablib==0.12.1
unicodecsv==0.14.1
whitenoise==3.3.1
xlrd==1.1.0
xlwt==1.3.0
django-filter==2.0.0.dev1
django-crispy-forms==1.7.0

